refactor: extract ThemeSwitcher component
